person-detector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
ultralytics==8.0.196
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
hnswlib==0.7.0
numpy==1.24.3
Pillow==10.1.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
scipy==1.11.4
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1